Create code.R
The R code for execution. The "all_stocks_1yrdata.csv" file from the link https://www.kaggle.com/code/kp4920/s-p-500-stock-data-time-series-analysis/data is used for data input in this file. So do download the data set and save it in the directory having the R file in the system before running this code.

Thank you.